+++ Advanced/negotiation-cordapp/README.md
@@ -51,7 +51,7 @@ First, go the the shell of PartyA, and propose a deal with yourself as buyer and
 
 We can now look at the proposals in the PartyA's vault:
 
-    run vaultQuery contractStateType: negotiation.states.ProposalState
+    run vaultQuery contractStateType: net.corda.samples.negotiation.states.ProposalState
 
 If we note down the state's `linearId.id`, we can now modify the proposal from the shell of PartyB by running:
 
@@ -64,4 +64,4 @@ Finally, let's have PartyA accept the proposal:
 We can now see the accepted trade in our vault with the new value by running the command (note we are now querying for
 `TradeState`s, not `ProposalState`s):
 
-    run vaultQuery contractStateType: negotiation.states.TradeState
+    run vaultQuery contractStateType: net.corda.samples.negotiation.states.TradeState
