+++ Advanced/negotiation-cordapp/gradle.properties
@@ -1,4 +1,4 @@
 name=Test
 group=com.negotiation
-version=0.2
+version=0.1
 kotlin.incremental=false
