+++ Advanced/negotiation-cordapp/settings.gradle
@@ -1,2 +1,2 @@
-include 'workflows-kotlin'
-include 'contracts-kotlin'
+include 'workflows'
+include 'contracts'
