+++ Advanced/negotiation-cordapp/workflows/src/main/kotlin/net/corda/samples/negotiation/flows/ModificationFlow.kt
@@ -1,0 +1,74 @@
+package net.corda.samples.negotiation.flows
+
+import co.paralleluniverse.fibers.Suspendable
+import net.corda.samples.negotiation.contracts.ProposalAndTradeContract
+import net.corda.samples.negotiation.states.ProposalState
+import net.corda.core.contracts.Command
+import net.corda.core.contracts.UniqueIdentifier
+import net.corda.core.flows.*
+import net.corda.core.node.services.queryBy
+import net.corda.core.node.services.vault.QueryCriteria
+import net.corda.core.transactions.SignedTransaction
+import net.corda.core.transactions.TransactionBuilder
+import net.corda.core.utilities.ProgressTracker
+
+object ModificationFlow {
+    @InitiatingFlow
+    @StartableByRPC
+    class Initiator(val proposalId: UniqueIdentifier, val newAmount: Int) : FlowLogic<Unit>() {
+        override val progressTracker = ProgressTracker()
+
+        @Suspendable
+        override fun call() {
+            // Retrieving the input from the vault.
+            val inputCriteria = QueryCriteria.LinearStateQueryCriteria(linearId = listOf(proposalId))
+            val inputStateAndRef = serviceHub.vaultService.queryBy<ProposalState>(inputCriteria).states.single()
+            val input = inputStateAndRef.state.data
+
+            // Creating the output.
+            val counterparty = if (ourIdentity == input.proposer) input.proposee else input.proposer
+            val output = input.copy(amount = newAmount, proposer = ourIdentity, proposee = counterparty)
+
+            // Creating the command.
+            val requiredSigners = listOf(input.proposer.owningKey, input.proposee.owningKey)
+            val command = Command(ProposalAndTradeContract.Commands.Modify(), requiredSigners)
+
+            // Building the transaction.
+            val notary = inputStateAndRef.state.notary
+            val txBuilder = TransactionBuilder(notary)
+            txBuilder.addInputState(inputStateAndRef)
+            txBuilder.addOutputState(output, ProposalAndTradeContract.ID)
+            txBuilder.addCommand(command)
+
+            // Signing the transaction ourselves.
+            val partStx = serviceHub.signInitialTransaction(txBuilder)
+
+            // Gathering the counterparty's signature.
+            val counterpartySession = initiateFlow(counterparty)
+            val fullyStx = subFlow(CollectSignaturesFlow(partStx, listOf(counterpartySession)))
+
+            // Finalising the transaction.
+            subFlow(FinalityFlow(fullyStx, listOf(counterpartySession)))
+        }
+    }
+
+    @InitiatedBy(Initiator::class)
+    class Responder(val counterpartySession: FlowSession) : FlowLogic<Unit>() {
+        @Suspendable
+        override fun call() {
+            val signTransactionFlow = object : SignTransactionFlow(counterpartySession) {
+                override fun checkTransaction(stx: SignedTransaction) {
+                    val ledgerTx = stx.toLedgerTransaction(serviceHub, false)
+                    val proposee = ledgerTx.inputsOfType<ProposalState>().single().proposee
+                    if (proposee != counterpartySession.counterparty) {
+                        throw FlowException("Only the proposee can modify a proposal.")
+                    }
+                }
+            }
+
+            val txId = subFlow(signTransactionFlow).id
+
+            subFlow(ReceiveFinalityFlow(counterpartySession, txId))
+        }
+    }
+}
